+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Unreleased
 
-- [added] Exposed the `CollectionReference` type from the `admin.firestore`
-  namespace.
+- [added] Exposed the `CollectionReference`, `WriteBatch`, `WriteResult` and
+  `QueryDocumentSnapshot` types from the `admin.firestore` namespace.
 
 # v6.0.0